Move template model and data to separate project
