+++ gradle/base/code-generator-model/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    `kotlin-dsl`
+}
+
+group = "junitbuild.base"
+
+repositories {
+    gradlePluginPortal()
+}
+
+dependencies {
+    api("gg.jte:jte:3.1.12")
+}
