+++ gradle/base/code-generator-model/src/main/resources/jre.yaml
