+++ gradle/base/settings.gradle.kts
@@ -1,0 +1,11 @@
+dependencyResolutionManagement {
+	versionCatalogs {
+		create("libs") {
+			from(files("../libs.versions.toml"))
+		}
+	}
+}
+
+rootProject.name = "base"
+
+include("code-generator-model")
