+++ gradle/plugins/code-generator/build.gradle.kts
@@ -11,4 +11,5 @@ dependencies {
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.17.1")
     implementation("gg.jte:jte:3.1.12")
     implementation(libs.classgraph)
+    implementation("junitbuild.base:code-generator-model")
 }
