+++ gradle/plugins/code-generator/src/main/resources/jre-templates/junit-jupiter-api/main/org/junit/jupiter/api/condition/JRE.java.jte
@@ -1,6 +1,6 @@
 @import java.util.List
 @import gg.jte.support.ForSupport
-@import junitbuild.generator.JRE
+@import junitbuild.generator.model.JRE
 
 @param List<JRE> jres
 @param String licenseHeader
