+++ gradle/plugins/code-generator/src/main/resources/jre-templates/junit-jupiter-engine/test/org/junit/jupiter/api/condition/EnabledOnJreIntegrationTests.java.jte
@@ -1,5 +1,5 @@
 @import java.util.List
-@import junitbuild.generator.JRE
+@import junitbuild.generator.model.JRE
 
 @param List<JRE> jres
 @param String licenseHeader
